requirements.txt
python-telegram-bot
langchain==0.3.18
langchain-community   
langchain-gigachat 
langchain-core
sentence-transformers 
faiss-cpu
datasets
numpy
scikit-learn
python-dotenv
pandas
torch
transformers
tika
psycopg2-binary